style: restructure, added services folder
